Enemy and bullets will now be killed upon collision.
For this the nodename is set to "dead" which will trigger a kill in the update.

To be able to test the nodename in the update an node object was added to the gameobject interface. So every gameobject has now a GENERIC node. Depending on the implementation of the game object and the exact needed nodetype, some manual typecasting needs to be done. This is already done for the existing classes.